Fix displayed name for GNOME Terminal and Black Box
libmacchina uses /proc/<pid>/comm, which is limited to 15 chars and
it's not guaranteed to hold a user friendly name.

For example GNOME Terminal is actually /usr/libexec/gnome-terminal-server
and its `comm` is `gnome-terminal-`.

Add a replacement function that returns user friendly names and
that can be expanded with more.